PERA-3130 Remove unnecessary json serializer
