+++ app/src/main/kotlin/com/algorand/android/modules/peraserializer/PeraSerializer.kt
@@ -28,12 +28,4 @@ class PeraSerializer @Inject constructor(val gson: Gson) : JsonSerializer {
             null
         }
     }
-
-    override fun <T> fromJson(json: Any, type: Class<T>): T? {
-        return try {
-            gson.fromJson(gson.toJson(json), type)
-        } catch (_: Exception) {
-            null
-        }
-    }
 }
