--- app/src/main/kotlin/com/algorand/android/ui/webview/bridge/serializer/WebViewJsonSerializer.kt
@@ -1,18 +1,0 @@
-/*
- * Copyright 2022-2025 Pera Wallet, LDA
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License
- */
-
-package com.algorand.android.ui.webview.bridge.serializer
-
-import com.algorand.android.modules.peraserializer.JsonSerializer
-import com.algorand.android.modules.peraserializer.PeraSerializer
-
-internal class WebViewJsonSerializer(jsonSerializerImpl: PeraSerializer) : JsonSerializer by jsonSerializerImpl
