+++ common-sdk/src/main/kotlin/com/algorand/wallet/foundation/json/JsonSerializer.kt
@@ -15,4 +15,5 @@ package com.algorand.wallet.foundation.json
 interface JsonSerializer {
     fun toJson(payload: Any?): String
     fun <T> fromJson(json: String, type: Class<T>): T?
+    fun <T> fromJson(json: Any, type: Class<T>): T?
 }
